Fixed known issue with key names
The first instances of keys with 'second' as their firstTimeSimple variable used to be joined in with the first set of Set A+SetB properties. It has been fixed to join the second set of SetA+SetB properties. New known issue: Pictures on search results are complemented based on position on screen, not based on actual layout type (paragraph, list, tab)